Colocando tudo na Main
