+++ modulo6/Rodada-cases-semana3/Cubo-challenge/front/cubo-challenge/src/Pages/Main/Main.jsx
@@ -1,47 +1,108 @@
 import React, { useEffect, useState } from "react";
 import axios from "axios";
-import { ChartSide, MainContainer, TableSide, ChartContainer } from './styled'
+import { ChartSide, PageContainer, HeaderContainer, InputContainer , MainContainer, TableSide, ChartContainer } from './styled'
 import { Chart as ChartJS, ArcElement, Tooltip, Legend } from 'chart.js';
 import { Doughnut } from 'react-chartjs-2';
 import {BASE_URL} from '../../Constants/base_url'
+import useForm from "../../Hooks/useForm";
 
 ChartJS.register(
     ArcElement, Tooltip, Legend
 )
 
-const Main = ({getParticipations, allParticipations}) => {
-    
-    
+const Main = () => {
+
     const [values, setValues] = useState([])
     const [names, setNames] = useState([])
+    const [allParticipations, setAllParticipations] = useState([])
+    const [form, onChange, clear] = useForm({first_name: "", last_name: "", participation: "" });
+
+
+
+    // GET para preencher a tabela
+    const getParticipations = async () => {
+    await axios
+    .get(BASE_URL)
+    .then((res) => {
+            setAllParticipations(res.data)
+        })
+        .catch((err) => {
+            console.log("Erro do Get", err.response)
+        })
+    }
+    
+
+    const createParticipation = async () => {
+        await axios
+        .post(BASE_URL, form)
+        .then((res) => {
+            getParticipations()
+            clear();
+        })
+        .catch((err) => {
+            alert(err.response.data);
+        })
+    }
+    
+
+    const onSubmitForm = (event) => {
+        event.preventDefault();
+        createParticipation();
+      };
+    
+    
+
+
+
+
+  
     
     //Doughnut
-        const data = {
-            datasets: [{
-                data: values,
-                backgroundColor: [
-                    'rgb(255, 99, 132)',
-                    'rgb(54, 162, 235)',
-                    'rgb(255, 205, 86)',
-                    'rgb(0, 224, 138)',
-                    'rgb(167, 32, 234)',
-                ],
-                borderColor: [
-                    'rgb(255, 99, 132)',
-                    'rgb(54, 162, 235)',
-                    'rgb(255, 205, 86)',
-                    'rgb(0, 224, 138)',
-                    'rgb(167, 32, 234)'
-                ],
-                borderWidth: 1,
+    ChartJS.register(
+        ArcElement, Tooltip, Legend
+    )
+    const data = {
+        datasets: [{
+            data: values,
+            backgroundColor: [
+                'rgb(255, 99, 132)',
+                'rgb(54, 162, 235)',
+                'rgb(255, 205, 86)',
+                'rgb(0, 224, 138)',
+                'rgb(167, 32, 234)',
+            ],
+            borderColor: [
+                'rgb(255, 99, 132)',
+                'rgb(54, 162, 235)',
+                'rgb(255, 205, 86)',
+                'rgb(0, 224, 138)',
+                'rgb(167, 32, 234)'
+            ],
+            borderWidth: 1,
 
-            }],
-            
+        }],
+        labels: names
+    };
 
-        
-            // These labels appear in the legend and in the tooltips when hovering different arcs
-            labels: names
-        };
+
+    // useEffect para renderizar na tela os ítens da requisição GET
+    useEffect(()=> {
+        getParticipations()
+    }, [])
+    
+    useEffect(()=>{
+        const userParticipation = allParticipations.map((value ) => {
+            return ((value.participation*100)/totalSum()).toFixed(1)
+        })
+        setValues(userParticipation)
+
+        const userNames = allParticipations.map((value ) => {
+            return value.first_name
+        })
+        setNames(userNames)
+    }, [allParticipations])
+
+
         
         //Map para pegar o valor total
         const mapPart = allParticipations.map((eachvalue)=>{
@@ -57,6 +118,7 @@ const Main = ({getParticipations, allParticipations}) => {
             }
             return sum
         }
+        
         // Map para pegar cada valor da tabela
         const mapParticipations = allParticipations.map((value)=>{
             return <tr key={value.id}>
@@ -68,53 +130,78 @@ const Main = ({getParticipations, allParticipations}) => {
     })
     
 
-    // useEffect para renderizar na tela os ítens da requisição GET
-        useEffect(() => {
-            getParticipations()
-        }, [])
-
-        useEffect(()=>{
-            const userParticipation = allParticipations.map((value ) => {
-                return ((value.participation*100)/totalSum()).toFixed(1)
-            })
-            setValues(userParticipation)
-        }, [allParticipations])
-
-        useEffect(()=>{
-            const userNames = allParticipations.map((value ) => {
-                return value.first_name
-            })
-            setNames(userNames)
-        }, [allParticipations])
+    
 
 
 
     
     return (
-        <MainContainer>
+        <PageContainer>
             
-            <TableSide>
-                <table>
-                    <thead>
-                        <tr>
-                            <th></th>
-                            <th>First name</th>
-                            <th>Last name</th>
-                            <th>Participation</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        {mapParticipations}
-                    </tbody>
-                </table>
-            </TableSide>
+            <HeaderContainer>
+                <InputContainer>
+                    <form onSubmit={onSubmitForm}>
+                        <input
+                            label="First Name"
+                            name='first_name'
+                            placeholder='First name'
+                            required
+                            type={"text"}
+                            value={form.first_name}
+                            onChange={onChange}
+                        />
 
-            <ChartSide>
-                <ChartContainer>
-                    <Doughnut data={data}/>
-                </ChartContainer>
-            </ChartSide>
-        </MainContainer>
+                        <input
+                            label="Last Name"
+                            name='last_name'
+                            placeholder='Last name'
+                            required
+                            type={"text"}
+                            value={form.last_name}
+                            onChange={onChange}
+                        />
+
+                        <input
+                            label="Participation"
+                            name='participation'
+                            placeholder='Participation'
+                            required
+                            type={"text"}
+                            value={form.participation}
+                            onChange={onChange}
+                        />
+
+                        <button type="submit">SEND</button>
+                    </form>
+                </InputContainer>
+            </HeaderContainer>
+
+            <MainContainer>
+                <TableSide>
+                                <table>
+                                    <thead>
+                                        <tr>
+                                            <th></th>
+                                            <th>First name</th>
+                                            <th>Last name</th>
+                                            <th>Participation</th>
+                                        </tr>
+                                    </thead>
+                                    <tbody>
+                                        {/* {mapParticipations.length > 0 ? mapParticipations : "Loading ..."} */}
+                                        {mapParticipations}
+                                    </tbody>
+                                </table>
+                            </TableSide>
+
+                            <ChartSide>
+                                <ChartContainer>
+                                    <Doughnut data={data}/>
+                                </ChartContainer>
+                            </ChartSide>
+            </MainContainer>
+            
+        </PageContainer>
     )
 }
